+++ irs-models/src/main/java/org/eclipse/tractusx/irs/component/Shell.java
@@ -23,54 +23,19 @@
  ********************************************************************************/
 package org.eclipse.tractusx.irs.component;
 
-import java.util.List;
-import java.util.Map;
-
-import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import com.fasterxml.jackson.databind.annotation.JsonPOJOBuilder;
-import io.swagger.v3.oas.annotations.media.Schema;
-import lombok.AllArgsConstructor;
 import lombok.Builder;
-import lombok.Singular;
-import lombok.Value;
+import org.eclipse.tractusx.irs.component.assetadministrationshell.AssetAdministrationShellDescriptor;
 
 /**
- * An AAS shell.
+ * Shell container
+ *
+ * @param contractAgreementId
+ * @param payload
  */
-@Value
-@Builder(toBuilder = true)
-@Schema(description = "")
-@AllArgsConstructor
-@JsonDeserialize(builder = Shell.ShellBuilder.class)
-public class Shell {
+@Builder
+public record Shell(String contractAgreementId, AssetAdministrationShellDescriptor payload) {
 
-    @Schema(implementation = String.class)
-    private String identification;
-
-    @Schema(implementation = String.class)
-    private String idShort;
-
-    @Schema()
-    @Singular
-    private Map<String, String> specificAssetIds;
-
-    @Schema()
-    @Singular
-    private List<Description> descriptions;
-
-    @Schema()
-    @Singular
-    private List<GlobalAssetIdentification> globalAssetIds;
-
-    @Schema()
-    @Singular
-    private List<SubmodelDescriptor> submodelDescriptors;
-
-    /**
-     * User to build Shell
-     */
-    @Schema(description = "User to build shell items")
-    @JsonPOJOBuilder(withPrefix = "")
-    public static class ShellBuilder {
+    public Shell withoutContractAgreementId() {
+        return Shell.builder().payload(this.payload()).build();
     }
 }